HandGunとSniperRifleを統合
