+++ Assets/EffectBullet.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 14a43162fdf254f4aa843cc9dbf5f274
-timeCreated: 1510669110
+guid: d1e56989fc766d44cbb1cbb0d9db813b
+timeCreated: 1510750530
 licenseType: Free
 MonoImporter:
   externalObjects: {}
