+++ Assets/SSP/Prefabs/Weapons/HandGun.prefab
@@ -29,24 +29,9 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!1 &1684454043637744
-GameObject:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  serializedVersion: 5
-  m_Component:
-  - component: {fileID: 4451773248574068}
-  m_Layer: 0
-  m_Name: Scope
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
 --- !u!1 &1722006071851656
 GameObject:
-  m_ObjectHideFlags: 1
+  m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
@@ -72,9 +57,9 @@ GameObject:
   - component: {fileID: 23128071150024078}
   - component: {fileID: 114397692148794256}
   - component: {fileID: 82332949709394078}
-  - component: {fileID: 114777560185994564}
-  - component: {fileID: 114930993466919184}
   - component: {fileID: 114611022078819548}
+  - component: {fileID: 114654731890088540}
+  - component: {fileID: 114802470638180542}
   m_Layer: 0
   m_Name: HandGun
   m_TagString: Untagged
@@ -88,12 +73,12 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1722006071851656}
-  m_LocalRotation: {x: 0, y: 1, z: 0, w: 0}
-  m_LocalPosition: {x: 0, y: 0, z: -0.73}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalRotation: {x: -0, y: 1, z: -0, w: 0}
+  m_LocalPosition: {x: -0.01, y: 0.514, z: -8.114656}
+  m_LocalScale: {x: 1, y: 0.34955806, z: 7.0200777}
   m_Children: []
-  m_Father: {fileID: 4184204856146398}
-  m_RootOrder: 0
+  m_Father: {fileID: 4515867202828200}
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
 --- !u!4 &4184204856146398
 Transform:
@@ -104,24 +89,10 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -0.01, y: 0.514, z: -2.99}
   m_LocalScale: {x: 1, y: 0.3495581, z: 7.020078}
-  m_Children:
-  - {fileID: 4086660066666890}
+  m_Children: []
   m_Father: {fileID: 4515867202828200}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!4 &4451773248574068
-Transform:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1684454043637744}
-  m_LocalRotation: {x: 0, y: 1, z: 0, w: 0}
-  m_LocalPosition: {x: 0, y: 0.81, z: -2.04}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 4515867202828200}
-  m_RootOrder: 1
-  m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
 --- !u!4 &4515867202828200
 Transform:
   m_ObjectHideFlags: 1
@@ -133,7 +104,7 @@ Transform:
   m_LocalScale: {x: 0.029826352, y: 0.10867492, z: 0.027502982}
   m_Children:
   - {fileID: 4184204856146398}
-  - {fileID: 4451773248574068}
+  - {fileID: 4086660066666890}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -371,7 +342,7 @@ MonoBehaviour:
   model: {fileID: 0}
   canInteract: 0
   hand: 1
---- !u!114 &114777560185994564
+--- !u!114 &114654731890088540
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -379,34 +350,35 @@ MonoBehaviour:
   m_GameObject: {fileID: 1996856847404154}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f0b914f142c9ed5438ce56c2ebc9a5af, type: 3}
+  m_Script: {fileID: 11500000, guid: 51783ba17acb6f347b8f8cf5fe341720, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  model: {fileID: 114802470638180542}
+  muzzle: {fileID: 1722006071851656}
+  gunHoldOffset: {x: -0.24, y: 0.1, z: -0.15}
+  leftHandOffset: {x: 0.09, y: -0.13, z: -0.12}
+  socpeCameraOffset: {x: -0.1, y: 0.2, z: 0.1}
+  bullet: {fileID: 1610161927253904, guid: 0e328b3de743cd8468a377b107845ced, type: 2}
+--- !u!114 &114802470638180542
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1996856847404154}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 950741a1567c45348af8526ad0eb7838, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   name: HandGun
   type: 0
   damageAmount: 100
   isOwnerLocalPlayer: 0
-  image: {fileID: 21300000, guid: 1f3b5537aebf5f34f9cd93065de4970f, type: 3}
+  image: {fileID: 21300000, guid: 90cafd6355d4e624198a9b604136600d, type: 3}
   playerId: 0
   teamId: 0
   coolTime: 0.5
   scopeMagnification: 1.5
-  bulletVelocity: 10
-  bulletDeathTime: 5
---- !u!114 &114930993466919184
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1996856847404154}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d9281a803051c144f9754bf6cc400432, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  model: {fileID: 114777560185994564}
-  muzzle: {fileID: 1722006071851656}
-  gunHoldOffset: {x: -0.24, y: 0.1, z: -0.15}
-  leftHandOffset: {x: 0.09, y: -0.13, z: -0.12}
-  socpeCameraOffset: {x: -0.1, y: 0.2, z: 0.1}
-  bullet: {fileID: 1302414001687628, guid: d364dda7de5cd074fa642a49f756148f, type: 2}
+  bulletVelocity: 50
+  bulletDeathTime: 3
+  RayDistance: 100
