+++ Assets/SSP/Scripts/Weapons/SemiAutoRifle.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d9281a803051c144f9754bf6cc400432
-timeCreated: 1509260159
+guid: 51783ba17acb6f347b8f8cf5fe341720
+timeCreated: 1510749167
 licenseType: Free
 MonoImporter:
   externalObjects: {}
